[refactor] ProjectDetail Custom hook 추가
- Custom Hook을 이용한 Project 상태 관리 추가
  `isEditing`과 `titleInput`은 서로간, project상태와 의존성이 있기 때문에
  custom hook으로 빼지는 않음.
