+++ src/pages/ProjectDetail.tsx
@@ -8,6 +8,8 @@ import ProjectDetailHeader from '../components/ProjectDetail/ProjectDetailHeader
 import ProjectUserInfo from '../components/Projects/ProjectsUserInfo';
 import InviteMember from '../components/ProjectDetail/InviteMember';
 
+import { useProject } from './hooks/ProjectDetailHooks';
+
 export interface IUser {
   uid: number;
   nickname: string;
@@ -21,32 +23,18 @@ export interface IProject {
   users: IUser[];
 }
 
-const initialData: IProject = {
-  name: 'Panopticon',
-  description: '부스트캠프 11조 B팀 입니다.',
-  owner: { uid: 0, nickname: 'boostcamp', email: 'boostcamp.connect.or.kr' },
-  users: [
-    { uid: 1, nickname: 'juyoungpark718', email: 'junsushin4546@gmail.com' },
-    { uid: 2, nickname: 'junsushin-dev', email: 'junsushin-dev@gmail.com' },
-    { uid: 3, nickname: 'saeeng', email: 'saeeng@gmail.com' },
-    { uid: 4, nickname: 'EarlyHail', email: 'hojin5633@gmail.com' },
-  ],
-};
-
 interface MatchParams {
   id: string;
 }
 
 function ProjectDetail(): React.ReactElement {
-  const [isEditing, setIsEditing] = useState(false);
-  const [titleInput, setTitleInput] = useState('');
   const match = useRouteMatch<MatchParams>('/project/:id');
   const projectId = match?.params.id;
-  /**
-   * @TODO
-   * projectID로 데이터 받아오기
-   */
-  const [project, setProject] = useState(initialData);
+
+  const [project, setProjectName, setProjectUsers] = useProject(projectId as string);
+  const [isEditing, setIsEditing] = useState(false);
+  const [titleInput, setTitleInput] = useState('');
+
   const startEdit = () => {
     setTitleInput(() => project.name);
     setIsEditing(true);
@@ -56,31 +44,16 @@ function ProjectDetail(): React.ReactElement {
     setIsEditing(false);
   };
   const changeTitle = () => {
-    /**
-     * @TODO
-     * 서버에 project name 변경하기
-     */
-    setProject(() => {
-      const newProject = _.cloneDeep(project);
-      newProject.name = titleInput;
-      return newProject;
-    });
+    setProjectName(titleInput);
     endEdit();
   };
   const titleInputChange = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
     setTitleInput(() => target.value);
   };
   const deleteUsers = (selectedUids: number[]): void => {
-    setProject(() => {
-      const newProject = _.cloneDeep(project);
-      newProject.users = newProject.users.filter((user) => !selectedUids.includes(user.uid));
-      /**
-       * @TODO
-       * 서버로 API 요청보내서 사용자 정보 지우기
-       */
-      return newProject;
-    });
+    setProjectUsers(selectedUids);
   };
+
   return (
     <Box p={5} display="flex" flexDirection="column">
       <Box display="flex" flexDirection="row" alignItems="center">
